pyproject.toml
[tool.poetry]
name = "ocrworker"
version = "0.4"
description = "OCR Worker"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
ocrmypdf = "^16.11"
ocrmypdf-papermerge = "0.7.1"
celery = {extras = ["redis"], version = "^5.5.3"}
sqlalchemy = "^2.0.44"
typer = "^0.20"
pydantic = "^2.12.3"
pydantic-settings = "^2.11.0"
pyyaml = "^6.0.3"
pikepdf = "^9.11.0"
botocore = "^1.35.99"
boto3 = "^1.35.99"
httpx = "^0.28.1"
psycopg2 = { version = "^2.9.11", optional = true}
mysqlclient = {version = "^2.2.7", optional = true}
lxml = "^5.4.0"
opencv-python = "^4.10"
numpy = "^2.0"
pdf2image = "^1.17"
uuid-extensions = "^0.1"
pillow = "^11.0"

[tool.poetry.extras]
mysql = ["mysqlclient"]
pg = ["psycopg2"]
databases = ["mysqlclient", "psycopg2"]

[tool.poetry.scripts]
ocr = "ocrworker.cli.ocr:app"

[tool.taskipy.tasks]
worker = "celery -A ocrworker.celery_app worker"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pre-commit = "^4.3.0"
taskipy = "^1.12.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
